feat(ui): replace MiniPlayer with sticky PlaybackBar portal above BottomNav
- Extract AB loop UI into new LoopSection component
- Rewrite PlaybackBar to transport + seek only (5 props vs 22)
- Portal PlaybackBar into #playback-bar-portal div rendered by App when activeSong is set
- Remove MiniPlayer component and all related handlers from App
- Update index.css: replace .mini-player* rules with #playback-bar-portal fixed styles
- Update --mini-player-height -> --transport-bar-height (96px)
- All 359 frontend tests pass; backend coverage 96%
